requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.18.0
requests>=2.31.0
beautifulsoup4>=4.12.0
scikit-learn>=1.3.0
sentence-transformers>=2.6.0
pyyaml>=6.0.0
ad-api>=1.0.2
python-dateutil>=2.8.0
